Added      ccweb-start/src/main/resources/lang.yml Modified   ccweb-core/src/main/java/ccait/ccweb/controllers/BaseController.java Modified   ccweb-core/src/main/java/ccait/ccweb/interceptor/SecurityInterceptor.java Modified   ccweb-core/src/main/java/ccait/ccweb/model/AclModel.java Modified   ccweb-core/src/main/java/ccait/ccweb/model/GroupModel.java Modified   ccweb-core/src/main/java/ccait/ccweb/model/PrivilegeModel.java Modified   ccweb-core/src/main/java/ccait/ccweb/model/QueryInfo.java Modified   ccweb-core/src/main/java/ccait/ccweb/model/RoleModel.java Modified   ccweb-core/src/main/java/ccait/ccweb/model/UserGroupRoleModel.java Modified   ccweb-core/src/main/java/ccait/ccweb/model/UserModel.java Modified   ccweb-core/src/main/java/ccait/ccweb/trigger/DefaultTrigger.java Modified   ccweb-core/src/main/java/ccait/ccweb/trigger/UserTableTrigger.java Modified   ccweb-start/pom.xml Modified   release/ccweb-start-1.0.0-SNAPSHOT.jar Modified   release/libs/entity.queryable-2.0-SNAPSHOT.jar

+++ ccweb-core/src/main/java/ccait/ccweb/model/AclModel.java
@@ -67,11 +67,11 @@ public class AclModel extends Queryable<AclModel> {
     this.modifyOn = modifyOn;
   }
 
-  public Long getModifyBy() {
+  public Integer getModifyBy() {
     return modifyBy;
   }
 
-  public void setModifyBy(Long modifyBy) {
+  public void setModifyBy(Integer modifyBy) {
     this.modifyBy = modifyBy;
   }
 
@@ -88,7 +88,7 @@ public class AclModel extends Queryable<AclModel> {
   private Date modifyOn;
 
   @Fieldname("${entity.table.reservedField.modifyBy:modifyBy}")
-  private Long modifyBy;
+  private Integer modifyBy;
 
   public UUID getAclId() {
     return this.aclId;
